pyproject.toml
[tool.poetry]
name = "ag-case"
version = "0.1.0"
description = ""
authors = ["nestor <albquenes7294@gmail.com>"]
readme = "README.md"
packages = [{include = "ag_case"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.9"
getenv = "^0.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
fastapi-mail = "^1.2.7"
pyjwt = "^2.6.0"
bcrypt = "^4.0.1"
secure-smtplib = "^0.1.1"
python-multipart = "^0.0.6"
pymupdf = "^1.21.1"
pypdf2 = "^3.0.1"
pdfminer-six = "^20221105"
pdf2image = "^1.16.3"
numpy = "^1.24.2"
opencv-python = "^4.7.0.72"
matplotlib = "^3.7.1"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
